core/standards/fullscreen/interactive/dom-fullscreen-enabled-property.html

<!DOCTYPE html>
<html>
<head>
	<title>fullscreenEnabled property</title>
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script>
		var test = async_test("fullscreenEnabled property");

		onload = function()
		{
			document.querySelector('button').addEventListener('click', goFullscreen, false);
		}

		function goFullscreen()
		{
			test.step(function() {
				assert_equals(document.fullscreenEnabled, true, "Property set when in fullscreen");
			})

			this.requestFullscreen();

			test.step(function() {
				assert_equals(document.fullscreenEnabled, true, "Property still set when in fullscreen");
			})

			test.done();
			removeFullscreenElement();
		}

		function removeFullscreenElement()
		{
			var el = document.fullscreenElement;
			el.parentNode.removeChild(el);
		}
	</script>
</head>
<body>

	<p>Click box below to trigger fullscreen mode.</p>

	<button>enter fullscreen</button>

	<div id="log"></div>

</body>
</html>
